All issues are resolved. Entities are replaced with substitions that can wait until the final product is built.
A less unnatural, but rather complicated fix for the namespace issue was
implemented: regex with nested groups and replacement functions, and that's
just the start!

Replacing ElementTree with a less fussy XML library might be indicated.

New issue! Makefile rules are not working for the raw to mml step.
Could be an old issue, one supposes. Tests are not implemented.