Untracked build targets, added a zip archive of them instead. They are nice to provide, being text and useful stand-alone, but it gets messy fast. Hence the usual practice of not tracking these :) A script to refresh the archive is provided.
Fixed a small omission with glaring consequences when making the "m" files:
entities were not substituted, and those symbols came out as big ugly text,
and the layout was thrown way out of whack!

The makefile is getting bigger and somewhat better organized. Phony targets,
really, as the actual file processing has been in place a while. Still not
working for the src/ raw files with intermediates/ mml targets.

Keeping the README current.