Fixed an error in the streamlining from the last commit. Passes the tests!
Tweaks to reST, big additions to comments/docs of ``lib/parameters``.

List of dependencies has links.

lib/parameters has some entries that are fixed, but don't really warrant a
standalone file and shouldn't be hard coded into the Python source. One for
each "root dependency" was defined, with the other parameters for that
dependency. The sensible thing was done by moving them to a standalone group
at the end of the file with barriers and warnings nearby.